+++ src/pages/OptionalHolidays.js
@@ -37,6 +37,8 @@ const OptionalHolidays = () =>
         <ul>
           <li><a href="/provinces/AB">Alberta holidays</a></li>
           <li><a href="/provinces/BC">British Columbia holidays</a></li>
+          <li><a href="/provinces/MB">Manitoba holidays</a></li>
+          <li><a href="/provinces/MB">Ontario holidays</a></li>
         </ul>
 
         <h2>How are statutory holidays different from optional holidays?</h2>