+++ src/utils/index.js
@@ -113,11 +113,14 @@ const checkRedirectYear = (req, res, next) => {
   next()
 }
 
-// middleware to add "optional=true" where provinceId is Alberta or BC, if no query param is set
+// middleware to add "optional=true" where provinceId is AB/BC/MB/ON, if no query param is set
 const optionalTrue = (req, res, next) => {
   const provinceId = req.params.provinceId
 
-  if ((provinceId === 'AB' || provinceId === 'BC') && req.query.optional === undefined) {
+  if (
+    (provinceId === 'AB' || provinceId === 'BC' || provinceId === 'MB' || provinceId === 'ON') &&
+    req.query.optional === undefined
+  ) {
     req.query.optional = 'true'
   }